Initialize Backend Router for TYPO3 8.7.49 ELTS (#75)
This is the backport of #47 for the 7-8 branch.

The monitoring of TYPO3 8.7 ELTS instances with the latest security updates (8.7.49) and the corresponding regression update ist not possible currently. This is due to the new ServerResponseCheck which makes use of the Backend Routes.

Can you push a new 1.0.x release please?